Create grouped and common prop type functionality.
