+++ .eslint.json
@@ -227,7 +227,7 @@
         "react/no-unknown-property": 2,
         "react/prefer-es6-class": [2, "always"],
         "react/prefer-stateless-function": 2,
-        "react/prop-types": 2,
+        "react/prop-types": 1,
         "react/react-in-jsx-scope": 2,
         "react/require-extension": 0,
         "react/self-closing-comp": 2,
