+++ src/components/Accordion/Item.js
@@ -4,13 +4,12 @@
  * @link        http://titon.io
  */
 
-/* eslints react/jsx-handler-names: 0 */
-
 import React, { PropTypes } from 'react';
 import Component from '../../Component';
 import Header from './Header';
 import Section from './Section';
 import collection from '../../prop-types/collection';
+import { showHidePropTypes } from '../propTypes';
 import CONTEXT_TYPES from './contextTypes';
 import MODULE from './module';
 
@@ -20,14 +19,11 @@ export default class Item extends Component {
     static contextTypes = CONTEXT_TYPES;
 
     static propTypes = {
+        ...showHidePropTypes,
         children: PropTypes.node,
         header: PropTypes.node.isRequired,
         index: PropTypes.number.isRequired,
-        onClickHeader: collection.func,
-        onHidden: collection.func,
-        onHiding: collection.func,
-        onShowing: collection.func,
-        onShown: collection.func
+        onClickHeader: collection.func
     };
 
     /**
