+++ src/components/Drop/Drop.js
@@ -10,8 +10,8 @@ import Menu from './Menu';
 import Toggle from './Toggle';
 import bind from '../../decorators/bind';
 import children from '../../prop-types/children';
-import collection from '../../prop-types/collection';
 import isOutsideElement from '../../utility/isOutsideElement';
+import { showHidePropTypes } from '../propTypes';
 import CONTEXT_TYPES from './contextTypes';
 import MODULE from './module';
 
@@ -21,11 +21,8 @@ export default class Drop extends Component {
     static childContextTypes = CONTEXT_TYPES;
 
     static propTypes = {
-        children: children(Menu, Toggle),
-        onHidden: collection.func,
-        onHiding: collection.func,
-        onShowing: collection.func,
-        onShown: collection.func
+        ...showHidePropTypes,
+        children: children(Menu, Toggle)
     };
 
     state = {
