+++ src/components/Form/ActionBar.js
@@ -4,16 +4,12 @@
  * @link        http://titon.io
  */
 
-import React, { PropTypes } from 'react';
+import React from 'react';
 import Component from '../../Component';
 import MODULE from './module';
 
 export default class ActionBar extends Component {
     static module = MODULE;
-
-    static propTypes = {
-        children: PropTypes.node
-    };
 
     /**
      * Render the form action bar.
