+++ src/components/Mask/Toggle.js
@@ -4,7 +4,7 @@
  * @link        http://titon.io
  */
 
-import React, { PropTypes } from 'react';
+import React from 'react';
 import Component from '../../Component';
 import bind from '../../decorators/bind';
 import CONTEXT_TYPES from './contextTypes';
@@ -14,10 +14,6 @@ export default class Toggle extends Component {
     static module = MODULE;
 
     static contextTypes = CONTEXT_TYPES;
-
-    static propTypes = {
-        children: PropTypes.node
-    };
 
     /**
      * Handlers for toggling the display of the mask overlay.
