+++ src/components/Pagination/Spacer.js
@@ -4,7 +4,7 @@
  * @link        http://titon.io
  */
 
-import React, { PropTypes } from 'react';
+import React from 'react';
 import Component from '../../Component';
 import CONTEXT_TYPES from './contextTypes';
 import MODULE from './module';
@@ -13,10 +13,6 @@ export default class Spacer extends Component {
     static module = MODULE;
 
     static contextTypes = CONTEXT_TYPES;
-
-    static propTypes = {
-        children: PropTypes.node
-    };
 
     /**
      * Render the pagination item spacer.
