+++ src/components/Tabs/Section.js
@@ -6,7 +6,7 @@
 
 import React, { PropTypes } from 'react';
 import Component from '../../Component';
-import collection from '../../prop-types/collection';
+import { showHidePropTypes } from '../propTypes';
 import CONTEXT_TYPES from './contextTypes';
 import MODULE from './module';
 
@@ -16,12 +16,9 @@ export default class Section extends Component {
     static contextTypes = CONTEXT_TYPES;
 
     static propTypes = {
+        ...showHidePropTypes,
         children: PropTypes.node,
-        index: PropTypes.number.isRequired,
-        onHidden: collection.func,
-        onHiding: collection.func,
-        onShowing: collection.func,
-        onShown: collection.func
+        index: PropTypes.number.isRequired
     };
 
     /**
